neocities: fix issue with deleting nonexistent files/dirs
Since fs.Rmdir requires an error to be returned if the directory to
delete does not actually exist, but object.Remove should not produce
errors if the same condition occurs, the previous strategy of
handling the error in apiDelete was inadequate. This has been fixed
to produce the correct / desired behaviour.